refactor Manager into partial classes
