+++ sdk/monitor/Azure.Monitor.OpenTelemetry.LiveMetrics/src/Internals/Manager.cs
@@ -2,14 +2,11 @@
 // Licensed under the MIT License.
 
 using System;
-using System.Diagnostics;
 using System.Threading;
 using Azure.Core.Pipeline;
-using Azure.Monitor.OpenTelemetry.Exporter.Internals;
 using Azure.Monitor.OpenTelemetry.Exporter.Internals.ConnectionString;
 using Azure.Monitor.OpenTelemetry.Exporter.Internals.Platform;
 using Azure.Monitor.OpenTelemetry.LiveMetrics.Internals.Diagnostics;
-using Azure.Monitor.OpenTelemetry.LiveMetrics.Models;
 
 namespace Azure.Monitor.OpenTelemetry.LiveMetrics.Internals
 {
@@ -19,13 +16,6 @@ namespace Azure.Monitor.OpenTelemetry.LiveMetrics.Internals
         private readonly ConnectionVars _connectionVars;
         private readonly bool _isAadEnabled;
         private readonly string _streamId = Guid.NewGuid().ToString();
-        private Timer _timer;
-        private string _etag = string.Empty;
-        private Action<object> _callbackAction = obj => { };
-
-        internal static bool? s_isAzureWebApp = null;
-
-        internal readonly State _state = new();
 
         public Manager(LiveMetricsExporterOptions options, IPlatform platform)
         {
@@ -86,143 +76,6 @@ namespace Azure.Monitor.OpenTelemetry.LiveMetrics.Internals
             }
 
             return new QuickPulseSDKClientAPIsRestClient(new ClientDiagnostics(options), pipeline, host: connectionVars.LiveEndpoint);
-        }
-
-        private void SetPingTimer()
-        {
-            _state.Update(LiveMetricsState.Ping);
-            _callbackAction = OnPing;
-            _timer.Change(dueTime: 0, period: 5000);
-        }
-
-        private void SetPostTimer()
-        {
-            _state.Update(LiveMetricsState.Post);
-            _callbackAction = OnPost;
-            _timer.Change(dueTime: 0, period: 1000);
-        }
-
-        private void OnCallback(object state) => _callbackAction.Invoke(state);
-
-        private void OnPing(object state)
-        {
-            try
-            {
-                Debug.WriteLine($"{DateTime.Now}: OnPing invoked.");
-
-                var response = _quickPulseSDKClientAPIsRestClient.PingCustom(
-                    ikey: _connectionVars.InstrumentationKey,
-                    apikey: null,
-                    xMsQpsTransmissionTime: null,
-                    xMsQpsMachineName: "Desktop-Name",
-                    xMsQpsInstanceName: "Desktop-Name",
-                    xMsQpsStreamId: _streamId,
-                    xMsQpsRoleName: null,
-                    xMsQpsInvariantVersion: "5",
-                    xMsQpsConfigurationEtag: _etag,
-                    monitoringDataPoint: null,
-                    cancellationToken: default);
-
-                if (response.GetRawResponse().Headers.TryGetValue("x-ms-qps-configuration-etag", out string? etagValue) && etagValue != _etag)
-                {
-                    Debug.WriteLine($"OnPing: updated etag: {etagValue}");
-                    _etag = etagValue;
-                }
-
-                if (response.GetRawResponse().Headers.TryGetValue("x-ms-qps-subscribed", out string? subscribedValue) && Convert.ToBoolean(subscribedValue))
-                {
-                    Debug.WriteLine($"OnPing: Subscribed: {subscribedValue}");
-                    SetPostTimer();
-                }
-            }
-            catch (Exception ex)
-            {
-                Debug.WriteLine(ex);
-            }
-        }
-
-        /// <summary>
-        /// Send data to LiveMetrics service.
-        /// </summary>
-        /// <param name="state"></param>
-        /// <remarks>
-        /// dataPoint.Metrics.Add(new MetricPoint { Name = "\\ApplicationInsights\\Requests/Sec", Value = 0, Weight = 1 });
-        /// dataPoint.Metrics.Add(new MetricPoint { Name = "\\ApplicationInsights\\Request Duration", Value = 0, Weight = 0 });
-        /// dataPoint.Metrics.Add(new MetricPoint { Name = "\\ApplicationInsights\\Requests Failed/Sec", Value = 0, Weight = 1 });
-        /// dataPoint.Metrics.Add(new MetricPoint { Name = "\\ApplicationInsights\\Requests Succeeded/Sec", Value = 0, Weight = 1 });
-        /// dataPoint.Metrics.Add(new MetricPoint { Name = "\\ApplicationInsights\\Dependency Calls/Sec", Value = 0, Weight = 1 });
-        /// dataPoint.Metrics.Add(new MetricPoint { Name = "\\ApplicationInsights\\Dependency Call Duration", Value = 0, Weight = 0 });
-        /// dataPoint.Metrics.Add(new MetricPoint { Name = "\\ApplicationInsights\\Dependency Calls Failed/Sec", Value = 0, Weight = 1 });
-        /// dataPoint.Metrics.Add(new MetricPoint { Name = "\\ApplicationInsights\\Dependency Calls Succeeded/Sec", Value = 0, Weight = 1 });
-        /// dataPoint.Metrics.Add(new MetricPoint { Name = "\\ApplicationInsights\\Exceptions/Sec", Value = 0, Weight = 1 });
-        /// dataPoint.Metrics.Add(new MetricPoint { Name = "\\Memory\\Committed Bytes", Value = 41372430336, Weight = 1 });
-        /// dataPoint.Metrics.Add(new MetricPoint { Name = "\\Processor(_Total)\\% Processor Time", Value = 14.1891f, Weight = 1 });.
-        /// </remarks>
-        private void OnPost(object state)
-        {
-            try
-            {
-                Debug.WriteLine($"{DateTime.Now}: OnPost invoked.");
-
-                var dataPoint = GetDataPoint();
-
-                var response = _quickPulseSDKClientAPIsRestClient.PostCustom(
-                    ikey: _connectionVars.InstrumentationKey,
-                    apikey: null,
-                    xMsQpsConfigurationEtag: _etag,
-                    xMsQpsTransmissionTime: null,
-                    monitoringDataPoints: new MonitoringDataPoint[] { dataPoint },
-                    cancellationToken: default);
-
-                if (response.GetRawResponse().Headers.TryGetValue("x-ms-qps-configuration-etag", out string? etagValue) && etagValue != _etag)
-                {
-                    Debug.WriteLine($"OnPost: updated etag: {etagValue}");
-                    _etag = etagValue;
-                }
-
-                if (response.GetRawResponse().Headers.TryGetValue("x-ms-qps-subscribed", out string? subscribedValue) && !Convert.ToBoolean(subscribedValue))
-                {
-                    Debug.WriteLine($"OnPost: Subscribed: {subscribedValue}");
-                    _etag = string.Empty;
-                    SetPingTimer();
-                }
-            }
-            catch (Exception ex)
-            {
-                Debug.WriteLine(ex);
-            }
-        }
-
-        /// <summary>
-        /// Searches for the environment variable specific to Azure Web App.
-        /// </summary>
-        /// <returns>Boolean, which is true if the current application is an Azure Web App.</returns>
-        internal static bool? IsWebAppRunningInAzure()
-        {
-            const string WebSiteEnvironmentVariable = "WEBSITE_SITE_NAME";
-            const string WebSiteIsolationEnvironmentVariable = "WEBSITE_ISOLATION";
-            const string WebSiteIsolationHyperV = "hyperv";
-
-            if (!s_isAzureWebApp.HasValue)
-            {
-                try
-                {
-                    // Presence of "WEBSITE_SITE_NAME" indicate web apps.
-                    // "WEBSITE_ISOLATION"!="hyperv" indicate premium containers. In this case, perf counters
-                    // can be read using regular mechanism and hence this method retuns false for
-                    // premium containers.
-                    // TODO: switch to platform. Not necessary for POC.
-                    s_isAzureWebApp = !string.IsNullOrEmpty(Environment.GetEnvironmentVariable(WebSiteEnvironmentVariable)) &&
-                                    Environment.GetEnvironmentVariable(WebSiteIsolationEnvironmentVariable) != WebSiteIsolationHyperV;
-                }
-                catch (Exception ex)
-                {
-                    LiveMetricsExporterEventSource.Log.AccessingEnvironmentVariableFailedWarning(WebSiteEnvironmentVariable, ex);
-                    return false;
-                }
-            }
-
-            return s_isAzureWebApp;
         }
     }
 }
